WIP: Replace VDSDataHandler by request handlers
